feat: fetch injury info for displaying all injuries in menu
