+++ backend/models/injuryDisplay.go
@@ -1,0 +1,10 @@
+package models
+
+type InjuryDisplay struct {
+	Type         string
+	GamePosition string
+	Team         string
+	JerseyNumber string
+	FirstName    string
+	LastName     string
+}
